requirements.txt
numpy>=1.21.0
scipy>=1.7.0
Pillow>=8.0.0
torch>=1.9.0
torchvision>=0.10.0
lpips>=0.1.4
tqdm>=4.62.0
